fix: App not opening after closing on macOs
