+++ src/handlers/cloud/index.ts
@@ -1,4 +1,4 @@
-import { BrowserWindow, ipcMain, shell } from 'electron'
+import { ipcMain, shell } from 'electron'
 import { CloudHandlers, RawScript, Script } from './types'
 import { RunInCloudStateMachine } from './states'
 import { basename, extname, isAbsolute, join } from 'path'
@@ -6,6 +6,7 @@ import { SCRIPTS_PATH } from '@/constants/workspace'
 import { getTempScriptName } from '@/script'
 import { rm, writeFile } from 'fs/promises'
 import { logError } from '@/utils/errors'
+import { browserWindowFromEvent } from '@/utils/electron'
 
 async function createTempFile(script: RawScript) {
   const tempFileName = getTempScriptName()
@@ -38,10 +39,11 @@ function toScriptFile(script: Script) {
   }
 }
 
-export function initialize(browserWindow: BrowserWindow) {
+export function initialize() {
   let stateMachine: RunInCloudStateMachine | null = null
 
-  ipcMain.handle(CloudHandlers.Run, async (_event, script: Script) => {
+  ipcMain.handle(CloudHandlers.Run, async (event, script: Script) => {
+    const browserWindow = browserWindowFromEvent(event)
     const file = await toScriptFile(script)
 
     const absolutePath = !isAbsolute(file.path)
