+++ src/handlers/index.ts
@@ -1,11 +1,7 @@
-import { BrowserWindow } from 'electron'
 import * as auth from './auth'
 import * as cloud from './cloud'
 
-interface Services {
-  browserWindow: BrowserWindow
-}
-export function initialize({ browserWindow }: Services) {
-  auth.initialize(browserWindow)
-  cloud.initialize(browserWindow)
+export function initialize() {
+  auth.initialize()
+  cloud.initialize()
 }
