+++ src/utils/electron.ts
@@ -1,6 +1,6 @@
 import { AddToastPayload } from '@/types/toast'
 import { platform, arch } from 'os'
-import { app, nativeImage, WebContents } from 'electron'
+import { app, BrowserWindow, nativeImage, WebContents } from 'electron'
 import net from 'net'
 import path from 'path'
 
@@ -69,4 +69,16 @@ export function getAppIcon(isDev: boolean) {
     : path.join(process.resourcesPath, 'icons', 'logo.png')
 
   return nativeImage.createFromPath(iconPath)
+}
+
+export const browserWindowFromEvent = (
+  event: Electron.IpcMainEvent | Electron.IpcMainInvokeEvent
+) => {
+  const browserWindow = BrowserWindow.fromWebContents(event.sender)
+
+  if (!browserWindow) {
+    throw new Error('failed to obtain browserWindow')
+  }
+
+  return browserWindow
 }
